added test support for Visual Studio C++ with ACML
git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1832 a6533b66-300f-0410-b061-b36de6c4fb2d

+++ tests/linspace_test.cpp
@@ -40,7 +40,7 @@ int main(void)
 
 	double from = 0;
 	double to = 2;
-	double points = 1;
+	int points = 1;
 	vec v = linspace(from, to, points);
 	cout << "Generating a vector of " << points << " points, from " << from << " to " << to << std::endl;
 	if (points != v.length())
